Move GetLevelFromSeverity to utility function in tests
